pyproject.toml
[project]
name = "runpod-toolkit"
version = "0.1.0"
description = "RunPod compute, storage, and vLLM serving toolkit"
requires-python = ">=3.11"
dependencies = [
    "runpod>=1.7.0",
    "boto3>=1.34.0",
    "paramiko>=3.4.0",
    "requests>=2.31.0",
    "typer[all]>=0.9.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "rich>=13.0.0",
]

[project.scripts]
cloud = "runpod_toolkit.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["runpod_toolkit"]
